requirements.txt
pillow
opencv-python
numpy
torch
torchvision
openvino>=2023.3.0
nncf>=2.7
